prd/requirements.txt
langchain-core>=0.3.0
langchain-google-genai>=2.0.0
langgraph>=0.2.0
pydantic>=2.8.0
psycopg[binary]==3.2.7
python-dotenv==1.0.1
pytest>=8.0.0
pytest-asyncio>=0.23.0
supabase>=2.3.0,<3

